requirements.txt
PyMySQL~=1.0.2
PyYAML~=5.4.1
sqlparse~=0.4.1